Refactoring 예시 (#223)
